==> 01_data_collection.ipynb <==
# Web scrapping

#### Imports

#dataframe
import pandas as pd
import numpy as np

#selenium
import selenium
from selenium.webdriver import Firefox
from selenium.webdriver.firefox.options import Options
from selenium.webdriver.common.by import By
import time

import re

import warnings
warnings.filterwarnings('ignore')

### 1. attribute

#get page with attributes
opts = Options()
opts.headless = True
browser = Firefox(options=opts)
path1 = 'https://www.larvalabs.com/cryptopunks/attributes'
browser.get(path1)

table = browser.find_element(By.XPATH,'/html/body/div[2]/div/div[3]/div/div[2]/table')

columns=[]

for element in table.find_elements(By.TAG_NAME, "th"):    
    columns.append(element.text)
    
print(columns)

table = browser.find_element(By.XPATH,'/html/body/div[2]/div/div[3]/div/div[2]/table/tbody')

results=[]

for element in table.find_elements(By.TAG_NAME, "td"):    
    results.append(element.text)

data = {}
for i in range(6):
    data[columns[i]]=results[i::6]

df = pd.DataFrame(data)
df.drop(columns=['More Examples'], inplace=True)
df.insert(0, 'id', range(0, len(df)))
df.head()

df.to_csv('attribute.csv')

### 2. punk

header_list = ['ranking','punk_image','punk_id','minscore','2nd','category_score','att_count_score','attributes','punk_skin','punk_type','total_score']
df = pd.read_csv('ranks.csv', names=header_list)
df = df[['ranking','punk_id','attributes','punk_skin','punk_type']]
df.head(3)

#change of format
pattern='\d{1,}'
search = []    
for values in df['punk_id']:
    search.append(re.search(pattern, values).group())

df['id'] = search

#category
conditions = [(df['punk_type'] == 'Alien'),
              (df['punk_type'] == 'Ape'),
              (df['punk_type'] == 'Zombie'),
              (df['punk_type'] == 'Female'),
              (df['punk_type'] == 'Male')]

values = [0, 1, 2, 3, 4]
df['type'] = np.select(conditions, values)

#skin
conditions = [(df['punk_skin'] == 'Alien'),
              (df['punk_skin'] == 'Ape'),
              (df['punk_skin'] == 'Zombie'),
              (df['punk_skin'] == 'Albino'),
              (df['punk_skin'] == 'Dark'),
              (df['punk_skin'] == 'Light'),
              (df['punk_skin'] == 'Mid')]

values = [0,1, 2, 3, 4, 5, 6]
df['skin'] = np.select(conditions, values)

df[['a','b']] = df['punk_id'].str.split("\"", 1, expand=True)
df[['url','d']] = df['b'].str.split("\"", 1, expand=True)
df.drop(columns=['a','b','d','punk_id'], axis=1, inplace=True)

len(df['id'].unique())

df.head(2)

df.to_csv('punk.csv')

### 3. punk_attribute_relation

#get page with ranking and punks
opts = Options()
opts.headless = True
browser = Firefox(options=opts)

columns=['attribute','punk_id']
print(columns)

results=[]

for i in range(1,1001):
    num = str(i)
    path = 'https://www.larvalabs.com/cryptopunks/details/'+num
    browser.get(path)
    table = browser.find_element(By.XPATH,'//*[@id="punkDetails"]/div[2]/div/div')
    
    for element in table.find_elements(By.TAG_NAME, "a"):
        results.append(element.text)
        results.append(num)
    time.sleep(0.1)
print('done')

data = {}
for i in range(2):
    data[columns[i]]=results[i::2]

df = pd.DataFrame(data)
df.head()

df.to_csv('punk_attributes_relation.csv')

df = pd.read_csv('punk_attributes_relation.csv')

#number of attributes
values = list(range(0, 87))
#conditions
conditions = [(df['attribute'] == 'Beanie'),(df['attribute'] == 'Choker'),
             (df['attribute'] == 'Pilot Helmet'),(df['attribute'] == 'Tiara'),
             (df['attribute'] == 'Orange Side'),(df['attribute'] == 'Buck Teeth'),
             (df['attribute'] == 'Welding Goggles'),(df['attribute'] == 'Pigtails'),
             (df['attribute'] == 'Pink With Hat'),(df['attribute'] == 'Top Hat'),
             (df['attribute'] == 'Spots'),(df['attribute'] == 'Rosy Cheeks'),
             (df['attribute'] == 'Blonde Short'),(df['attribute'] == 'Wild White Hair'),
             (df['attribute'] == 'Cowboy Hat'),(df['attribute'] == 'Wild Blonde'),
             (df['attribute'] == 'Straight Hair Blonde'),(df['attribute'] == 'Big Beard'),
             (df['attribute'] == 'Red Mohawk'),(df['attribute'] == 'Half Shaved'),
             (df['attribute'] == 'Blonde Bob'),(df['attribute'] == 'Vampire Hair'),
             (df['attribute'] == 'Clown Hair Green'),(df['attribute'] == 'Straight Hair Dark'),
             (df['attribute'] == 'Straight Hair'),(df['attribute'] == 'Silver Chain'),
             (df['attribute'] == 'Dark Hair'),(df['attribute'] == 'Purple Hair'),
             (df['attribute'] == 'Gold Chain'),(df['attribute'] == 'Medical Mask'),
             (df['attribute'] == 'Tassle Hat'),(df['attribute'] == 'Fedora'),
             (df['attribute'] == 'Police Cap'),(df['attribute'] == 'Clown Nose'),
             (df['attribute'] == 'Smile'),(df['attribute'] == 'Cap Forward'),
             (df['attribute'] == 'Hoodie'),(df['attribute'] == 'Front Beard Dark'),
             (df['attribute'] == 'Frown'),(df['attribute'] == 'Purple Eye Shadow'),   
             (df['attribute'] == 'Handlebars'),(df['attribute'] == 'Blue Eye Shadow'),
             (df['attribute'] == 'Green Eye Shadow'),(df['attribute'] == 'Vape'),
             (df['attribute'] == 'Front Beard'),(df['attribute'] == 'Chinstrap'),
             (df['attribute'] == '3D Glasses'),(df['attribute'] == 'Luxurious Beard'),
             (df['attribute'] == 'Mustache'),(df['attribute'] == 'Normal Beard Black'),
             (df['attribute'] == 'Normal Beard'),(df['attribute'] == 'Eye Mask'),
             (df['attribute'] == 'Goat'),(df['attribute'] == 'Do-rag'),
             (df['attribute'] == 'Shaved Head'),(df['attribute'] == 'Muttonchops'),
             (df['attribute'] == 'Peak Spike'),(df['attribute'] == 'Pipe'),
             (df['attribute'] == 'VR'),(df['attribute'] == 'Cap'),
             (df['attribute'] == 'Small Shades'),(df['attribute'] == 'Clown Eyes Green'),
             (df['attribute'] == 'Clown Eyes Blue'),(df['attribute'] == 'Headband'),
             (df['attribute'] == 'Crazy Hair'),(df['attribute'] == 'Knitted Cap'),
             (df['attribute'] == 'Mohawk Dark'),(df['attribute'] == 'Mohawk'),
             (df['attribute'] == 'Mohawk Thin'),(df['attribute'] == 'Frumpy Hair'),
             (df['attribute'] == 'Wild Hair'),(df['attribute'] == 'Messy Hair'),
             (df['attribute'] == 'Eye Patch'),(df['attribute'] == 'Stringy Hair'),
             (df['attribute'] == 'Bandana'),(df['attribute'] == 'Classic Shadesv'),
             (df['attribute'] == 'Shadow Beard'),(df['attribute'] == 'Regular Shades'),
             (df['attribute'] == 'Horned Rim Glasses'),(df['attribute'] == 'Big Shades'),
             (df['attribute'] == 'Nerd Glasses'),(df['attribute'] == 'Black Lipstick'),
             (df['attribute'] == 'Mole'),(df['attribute'] == 'Purple Lipstick'),
             (df['attribute'] == 'Hot Lipstick'),(df['attribute'] == 'Cigarette'),
             (df['attribute'] == 'Earring')]

df['att_id'] = np.select(conditions, values)

df.drop(columns=['Unnamed: 0'], axis=1, inplace=True)

len(df['punk_id'])

df.drop_duplicates(subset=None, keep='first', inplace=True)

len(df['punk_id'])

len(df['punk_id'].unique())

df.loc[df['punk_id'] == 0]

df.drop([27954, 27956], inplace=True)
df.loc[df['punk_id'] == 0]

df.to_csv('punk_attribute_relation.csv')

### 4. transaction

#initiate browser
opts = Options()
opts.headless = True
browser = Firefox(options=opts)

results=[]

columns=['1','2','3','4','5','6','7','8','9','10']
print(columns)

#loop to fetch all transactions
for i in range(0,10000):
    num = str(i)
    path = 'https://www.larvalabs.com/cryptopunks/details/'+ num
    browser.get(path)
    table = browser.find_element(By.XPATH,'//*[@id="punkHistory"]/div/table/tbody')
    
    try:
        for element in table.find_elements(By.TAG_NAME, "td"):
            results.append(element.text)
            results.append(num)
    except:
        pass
    time.sleep(0.1) 

data = {}
for i in range(10):
    data[columns[i]]=results[i::10]

df = pd.DataFrame(data)
df = df.iloc[1: , :]
df.reset_index(inplace=True, drop=True)
df.head()

df.to_csv('transaction_all_data.csv')

df = pd.read_csv('transaction_all_data.csv')
df.head(3)

df.drop(columns=['Unnamed: 0','Unnamed: 0.1','2','3','4','5','6','8'], axis=1, inplace=True)
df.drop(df[df['1']=='Transfer'].index)
df = df.rename(columns={'10':'punk_id','1':'type'})
df['id'] = df.index
df.head(3)

df.isna().sum()

df['7'] = df['7'].fillna('0.00Ξ')
df.isna().sum()

df.reset_index(drop=True)
df[['amount','b']] = df['7'].str.split("Ξ", 1, expand=True)
df.drop(columns=['b'], axis=1, inplace=True)

df.isna().sum()

#change of format
pattern='\d.{0,}'
search = []    
for values in df['amount']:
    try:
        search.append(re.search(pattern, values).group())
    except:
        search.append(values)
df['amount'] = search
df[34950:34955]

len(df)

boolean_findings1 = df['amount'].str.contains('K|M|B|T|Y')
total_occurence1 = boolean_findings1.sum()
total_occurence1

boolean_findings1

df['choice'] = boolean_findings1
df[34950:34955]

df = df[~df['amount'].str.contains("Y")]
df = df[~df['amount'].str.contains("Z")]
df = df[~df['amount'].str.contains("E")]

a = df.loc[df['choice'] == False]
b = df.loc[df['choice'] == True]

b.head(3)

b['amount']=b['amount'].str.replace(',','.')

b.head(3)

c = b[b['amount'].str.contains("K")]#thousand
d = b[b['amount'].str.contains("M")]#million
e = b[b['amount'].str.contains("B")]#billion
f = b[b['amount'].str.contains("T")]#trillion

c.tail(3)

c['amount']=c['amount'].str.replace('K','')
d['amount']=d['amount'].str.replace('M','')
e['amount']=e['amount'].str.replace('B','')
f['amount']=f['amount'].str.replace('T','')

c.tail(3)

c['amount']=c['amount'].astype(float)
d['amount']=d['amount'].astype(float)
e['amount']=e['amount'].astype(float)
f['amount']=f['amount'].astype(float)

c['amount'] = c['amount'] * 1000
d['amount'] = d['amount'] * 1000000
e['amount'] = e['amount'] * 1000000000
f['amount'] = f['amount'] * 1000000000000
c.tail(3)

df = pd.concat([a, c, d, e, f], ignore_index=True)
df = df.sort_values(by=['id'])
df.reset_index(drop=True)
df.tail()

df.isna().sum()

#change of format
pattern='\d{4}'
search = []    
for values in df['9']:
    try:
        search.append(re.search(pattern, values).group())
    except:
        search.append(re.search(pattern, values))
df['year'] = search
df.head(3)

#change of format
pattern='\d{2}'
search = []    
for values in df['9']:
    try:
        search.append(re.search(pattern, values).group())
    except:
        search.append(re.search(pattern, values))
df['day'] = search
df.head(3)

#change of format
pattern='\w{3}'
search = []    
for values in df['9']:
    try:
        search.append(re.search(pattern, values).group())
    except:
        search.append(re.search(pattern, values))
df['mon'] = search
df.head(3)

#category
conditions = [(df['mon'] == 'Jan'),
              (df['mon'] == 'Feb'),
              (df['mon'] == 'Mar'),
              (df['mon'] == 'Apr'),
              (df['mon'] == 'May'),
              (df['mon'] == 'Jun'),
              (df['mon'] == 'Jul'),
              (df['mon'] == 'Aug'),
              (df['mon'] == 'Sep'),
              (df['mon'] == 'Oct'),
              (df['mon'] == 'Nov'),
              (df['mon'] == 'Dec'),]

values = ['01', '02', '03', '04', '05','06','07','08','09','10','11','12']
df['month'] = np.select(conditions, values)
df.head(5)

df['date'] = df[['year', 'month', 'day']].apply(lambda x: '-'.join(x), axis=1)
df.head(3)

df.drop(columns=['7','9','year','day','mon','month','choice'], axis=1, inplace=True)
df.head(3)

df['amount'] = df['amount'].astype(float)

df.info()

df.isna().sum()

len(df['punk_id'].unique())

df.tail(3)

df_sort = df.sort_values(by=['punk_id'])
df_sort.tail(5)

#what's missing? no attributes
mylist = list(df_sort['punk_id'].unique())
l = []
def missing_num(mylist):
    for n in range(0,10000): 
        if n not in mylist: 
            l.append(n)
    if not l:
        print("Nothing is missing")
        
missing_num(mylist)

df.to_csv('transaction.csv')
