==> 02_database.ipynb <==
# PostgreSQL

import psycopg2
import pandas as pd

try:
    connection = psycopg2.connect(host="localhost", 
                                  port = 5432,
                                  database='mydb',
                                  user="postgres",
                                  password="postgres")
except:
    print("I am unable to connect to the database")

cursor = connection.cursor()

cursor.execute("DROP TABLE IF EXISTS attribute")

cursor.execute("""CREATE TABLE attribute
                (id INT PRIMARY KEY, 
                 attribute VARCHAR(255))""")
connection.commit()

cursor.execute("""SELECT table_name FROM information_schema.tables
       WHERE table_schema = 'public'""")
for table in cursor.fetchall():
    print(table)

df = pd.read_csv('attribute.csv')
df.head(3)

sql = "INSERT INTO attribute (id, attribute) VALUES (%s,%s)"

#tuple and than list
df['att'] = df[['id', 'Attribute']].apply(tuple, axis=1)
att = list(df['att'])

cursor.executemany(sql, att)
connection.commit()
print(cursor.rowcount, "record inserted")

cursor.close()
connection.close()

# MySQL

import pymysql
import pandas as pd

try:
    connection = pymysql.connect(host='localhost',
                                 user='root',
                                 password='mysql')
except:
    print("I am unable to connect to the database")

cursor = connection.cursor()
cursor.execute("DROP DATABASE IF EXISTS mydb")
cursor.execute("CREATE DATABASE mydb")
cursor.execute("USE mydb") 

cursor.execute("DROP TABLE IF EXISTS attribute")

cursor.execute("""CREATE TABLE attribute
                (id INT PRIMARY KEY, 
                 attribute VARCHAR(255))""")
cursor.execute("SHOW TABLES")
for x in cursor:
    print(x)

df = pd.read_csv('attribute.csv')

sql = "INSERT INTO attribute (id, attribute) VALUES (%s,%s)"

#tuple and than list
df['att'] = df[['id', 'Attribute']].apply(tuple, axis=1)
att = list(df['att'])

cursor.executemany(sql, att)
connection.commit()
print(cursor.rowcount, "record inserted")

cursor.execute("DROP TABLE IF EXISTS type")

cursor.execute("""CREATE TABLE type
                (id INT PRIMARY KEY, 
                 category VARCHAR(255))""")
cursor.execute("SHOW TABLES")
for x in cursor:
    print(x)

sql = "INSERT INTO type (id, category) VALUES (%s,%s)"

types = [(0, 'Alien'),(1, 'Ape'),(2, 'Zombie'),(3, 'Female'),(4, 'Male')]

cursor.executemany(sql, types)
connection.commit()
print(cursor.rowcount, "record inserted")

cursor.execute("DROP TABLE IF EXISTS skin")

cursor.execute("""CREATE TABLE skin
                (id INT PRIMARY KEY, 
                 skin VARCHAR(255))""")
cursor.execute("SHOW TABLES")
for x in cursor:
    print(x)

sql = "INSERT INTO skin (id, skin) VALUES (%s,%s)"

val = [(0, 'Alien'),(1, 'Ape'),(2, 'Zombie'),(3, 'Albino'),(4, 'Dark'),(5, 'Light'),(6, 'Mid')]

cursor.executemany(sql, val)
connection.commit()
print(cursor.rowcount, "record inserted")

cursor.execute("DROP TABLE IF EXISTS punk")

cursor.execute("""CREATE TABLE punk
                (id INT PRIMARY KEY, 
                 url VARCHAR(255),
                 rarest_rank INT,
                 type_id INT ,
                 skin_id INT )""")

cursor.execute("SHOW TABLES")
for x in cursor:
    print(x)

df = pd.read_csv('punk.csv')

sql = "INSERT INTO punk (id, url, rarest_rank, type_id, skin_id) VALUES (%s,%s,%s,%s,%s)"

#tuple and than list
df['punk_all'] = df[['id','url', 'ranking','type','skin']].apply(tuple, axis=1)
val = list(df['punk_all'])

cursor.executemany(sql, val)
connection.commit()
print(cursor.rowcount, "record inserted")

cursor.execute("""ALTER TABLE punk ADD FOREIGN KEY (type_id) REFERENCES type(id)""")
cursor.execute("""ALTER TABLE punk ADD FOREIGN KEY (skin_id) REFERENCES skin(id)""")

cursor.execute("DROP TABLE IF EXISTS punk_attribute_relation")

cursor.execute("""CREATE TABLE punk_attribute_relation
                (punk_id INT,
                 attribute_id INT)""")
cursor.execute("SHOW TABLES")
for x in cursor:
    print(x)

df = pd.read_csv('punk_attribute_relation.csv')

sql = "INSERT INTO punk_attribute_relation (punk_id, attribute_id) VALUES (%s,%s)"

#tuple and than list
df['punk_att_all'] = df[['punk_id','att_id']].apply(tuple, axis=1)
val = list(df['punk_att_all'])

cursor.executemany(sql, val)
connection.commit()
print(cursor.rowcount, "record inserted")

cursor.execute("""ALTER TABLE punk_attribute_relation ADD FOREIGN KEY (punk_id) REFERENCES punk(id) ON UPDATE CASCADE""")
cursor.execute("""ALTER TABLE punk_attribute_relation ADD FOREIGN KEY (attribute_id) REFERENCES attribute(id) ON UPDATE CASCADE""")

cursor.execute("DROP TABLE IF EXISTS transaction")

cursor.execute("""CREATE TABLE transaction
                (id INT PRIMARY KEY,
                 type VARCHAR(255),
                 date DATE,
                 amount FLOAT,
                 punk_id INT)""")
cursor.execute("SHOW TABLES")
for x in cursor:
    print(x)

df = pd.read_csv('transaction.csv')

sql = "INSERT INTO transaction (id, type, date, amount, punk_id) VALUES (%s,%s,%s,%s,%s)"

#tuple and than list
df['all'] = df[['id','type','date','amount','punk_id']].apply(tuple, axis=1)
val = list(df['all'])

cursor.executemany(sql, val)
connection.commit()
print(cursor.rowcount, "record inserted")

cursor.execute("""ALTER TABLE transaction ADD FOREIGN KEY (punk_id) REFERENCES punk(id) ON UPDATE CASCADE""")

cursor.close()
