==> 04_final_dataset.ipynb <==
# Dataset preparation

#### Imports

#sql
import mysql.connector
from mysql.connector import Error
import pymysql

#dataframe
import pandas as pd
import numpy as np

#visualization
import matplotlib.pyplot as plt
import seaborn as sns

import warnings
warnings.filterwarnings('ignore')

Features to find: 
Punk number,Skin, Type, Rarest rank, Total number of attributes, Last sold price, Number of sold transactions, Total in sold transactions, Number of transactions, Highest sold price, Difference between two last sold transactions, Current hold time.

connection = pymysql.connect(host='localhost',
                             user='root',
                             password='mysql',
                             db='mydb')

cursor = connection.cursor()
cursor.execute("USE mydb")

#### Queries to get all features

try:
    query = """SELECT id, skin_id, type_id, rarest_rank FROM punk;"""
    df1 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

df1.rename(columns={'id': 'punk_id'}, inplace=True)

df1.head(2)

try:
    query = """SELECT punk_id, attribute_id FROM punk_attribute_relation;"""
    df2 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

df2 = df2.groupby(['punk_id']).size()
df2 = df2.to_frame()
df2.columns = ['total_att']
df2['punk_id'] = df2.index
df2 = df2.reset_index(drop=True)
df2.head(2)

len(df2)

new_row = {'total_att':0, 'punk_id':281}
df2 = df2.append(new_row, ignore_index=True)
new_row = {'total_att':0, 'punk_id':510}
df2 = df2.append(new_row, ignore_index=True)
new_row = {'total_att':0, 'punk_id':641}
df2 = df2.append(new_row, ignore_index=True)
new_row = {'total_att':0, 'punk_id':741}
df2 = df2.append(new_row, ignore_index=True)
new_row = {'total_att':0, 'punk_id':1050}
df2 = df2.append(new_row, ignore_index=True)
new_row = {'total_att':0, 'punk_id':2204}
df2 = df2.append(new_row, ignore_index=True)
new_row = {'total_att':0, 'punk_id':3307}
df2 = df2.append(new_row, ignore_index=True)
new_row = {'total_att':0, 'punk_id':6487}
df2 = df2.append(new_row, ignore_index=True)
len(df2)

df2 = df2.sort_values(by=['punk_id'])
df2 = df2.reset_index(drop=True)
df2.head(3)

try:
    query = """SELECT date, amount, punk_id 
    FROM transaction
    WHERE type = 'Sold';"""
    df3 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

#keep only first item for each punk
df3 = df3.drop_duplicates(subset=['punk_id'], keep='first')
df3.rename(columns={'amount': 'last_price'}, inplace=True)
df3.drop(['date'], axis=1, inplace=True)
df3.head(5)

len(df3)

try:
    query = """SELECT date, amount, punk_id 
    FROM transaction
    WHERE type = 'Sold';"""
    df4 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

df4 = df4.groupby(['punk_id']).count()
df4['punk_id'] = df4.index
df4 = df4.reset_index(drop=True)
df4.drop(['date'], axis=1, inplace=True)
df4.rename(columns={'amount': 'num_sold_trans'}, inplace=True)
df4.head(5)

len(df4)

try:
    query = """SELECT date, amount, punk_id 
    FROM transaction
    WHERE type = 'Sold';"""
    df5 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

df5 = df5.groupby(['punk_id']).sum()
df5['punk_id'] = df5.index
df5 = df5.reset_index(drop=True)
df5.rename(columns={'amount': 'total_sold_trans'}, inplace=True)
df5.head(5)

len(df5)

try:
    query = """SELECT date, amount, punk_id 
    FROM transaction;"""
    df6 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

df6 = df6.groupby(['punk_id']).count()
df6['punk_id'] = df6.index
df6.drop(['date'], axis=1, inplace=True)
df6 = df6.reset_index(drop=True)
df6.rename(columns={'amount': 'num_trans'}, inplace=True)
df6.head(5)

try:
    query = """SELECT date, amount, punk_id 
    FROM transaction
    WHERE type = 'Sold';"""
    df7 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

df7 = df7.sort_values(["punk_id","amount"], ascending = (True, False))
df7.head(3)

df7 = df7.drop_duplicates(subset=['punk_id'], keep='first')
df7.rename(columns={'amount': 'highest_price'}, inplace=True)
df7.drop(['date'], axis=1, inplace=True)
df7.head(3)

try:
    query = """SELECT date, amount, punk_id 
    FROM transaction
    WHERE type = 'Sold';"""
    df8 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

df8.rename(columns={'amount': 'last_price'}, inplace=True)
df8.drop(['date'], axis=1, inplace=True)
df8.head(3)

#last price
df8a = df8.groupby('punk_id').head(1)
df8a = df8a.reset_index(drop=True)
#previous price
df8b = df8.groupby('punk_id').head(2)
df8b = df8b.drop_duplicates(subset=['punk_id'], keep='last')
df8b = df8b.reset_index(drop=True)
df8b.rename(columns={'last_price': 'previous_price'}, inplace=True)

df8 = df8a.copy()
df8['last_prices_diff'] = (df8a['last_price']) - (df8b['previous_price'])
df8.drop(['last_price'], axis=1, inplace=True)
df8.head(10)

try:
    query = """SELECT date, amount, punk_id 
    FROM transaction
    WHERE type = 'Sold';"""
    df9 = pd.read_sql(query,connection)
except Exception as e:
    cursor.close()
    print(str(e))

#last price
df9 = df9.groupby('punk_id').head(1)
df9 = df9.reset_index(drop=True)

df9.rename(columns={'amount': 'last_price','date': 'last_date'}, inplace=True)
df9.head(3)

import datetime
x = datetime.date(2022,1,10)
y = datetime.date(2017,6,23)
print('Total time of CryptoPunks project', x-y)

df9['date_diff'] = x - (df9['last_date'])
df9.drop(['last_price','last_date'], axis=1, inplace=True)
df9.head(3)

df9['date_diff'] = df9['date_diff'].dt.days
df9.head(3)

#### Combine to final dataframe

#punks and number of attributes
df_final = pd.merge(df1, df2, on=["punk_id"])

#last soldprice
df_final = df_final.merge(df3, how='left')
df_final['last_price'] = df_final['last_price'].fillna(0)

#number of sold transactions
df_final = df_final.merge(df4, how='left')
df_final['num_sold_trans'] = df_final['num_sold_trans'].fillna(0)

#total in sold transactions
df_final = df_final.merge(df5, how='left')
df_final['total_sold_trans'] = df_final['total_sold_trans'].fillna(0)

#number of transactions
df_final = df_final.merge(df6, how='left')
df_final['num_trans'] = df_final['num_trans'].fillna(0)

#highest price
df_final = df_final.merge(df7, how='left')
df_final['highest_price'] = df_final['highest_price'].fillna(0)

#difference in last prices
df_final = df_final.merge(df8, how='left')
df_final['last_prices_diff'] = df_final['last_prices_diff'].fillna(0)

#current hold time
df_final = df_final.merge(df9, how='left')
df_final['date_diff'] = df_final['date_diff'].fillna(1662)

df_final.head(3)

df_final['highest_price'].max()

df_final = df_final[df_final['last_price'] < 120]
df_final = df_final[df_final['highest_price'] < 120]

print(df_final['last_price'].max(),df_final['highest_price'].max())

df = df_final.loc[(df_final['type_id'] == 3) | (df_final['type_id'] == 4)]
df.head(10)

df.to_csv('final_dataset.csv')
